Add registration functionality

+++ base/views.py
@@ -1,8 +1,10 @@
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
 from django.contrib.auth.models import User
 from django.db.models import Q
+from django.http import HttpResponse
 from django.shortcuts import redirect, render
 
 from .forms import RoomForm
@@ -10,9 +12,31 @@ from .models import Room, Topic
 
 
 # Create your views here.
-def loginPage(request):
+def registerUser(request):
+    form = UserCreationForm()
+
     if request.method == "POST":
-        username = request.POST.get("username")
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect("home")
+        else:
+            messages.error(request, "An error occurred during registration.")
+
+    return render(request, "base/login_register.html", {"form": form})
+
+
+def loginPage(request):
+    page = "login"
+
+    if request.user.is_authenticated:
+        return redirect("home")
+
+    if request.method == "POST":
+        username = request.POST.get("username").lower()
         password = request.POST.get("password")
 
         try:
@@ -28,7 +52,7 @@ def loginPage(request):
         else:
             messages.error(request, "Username or password does not exist")
 
-    context = {}
+    context = {"page": page}
     return render(request, "base/login_register.html", context)
 
 
@@ -72,8 +96,13 @@ def createRoom(request):
     return render(request, "base/room_form.html", context)
 
 
+@login_required(login_url="login")
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!!")
+
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -85,8 +114,13 @@ def updateRoom(request, pk):
     return render(request, "base/room_form.html", context)
 
 
+@login_required(login_url="login")
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!!")
+
     if request.method == "POST":
         room.delete()
         return redirect("home")
